// FRONTEND_NO_NODE_MODULES/FRONTEND OF BACKEND/admin/lib/client/models/DealerResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DealerResponse = {
    id?: number;
    publicId?: string;
    code?: string;
    name?: string;
    email?: string;
    phone?: string;
    address?: string;
    creditLimit?: number;
    outstandingBalance?: number;
    portalEmail?: string;
    generatedPassword?: string;
};